Create Auto-reboot
This script automatically restarts the Web Server service of the SonicWall GMS, and that service restarts all of the other services. This script will also send out an email saying that the service is being restarted and then it will send a follow up email once it has recovered. 

Currently the script is configured to use the Google SMTP relay.